backend/requirements.txt
fastapi==0.115.1
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.11.4
pydantic-settings==2.2.1
pre-commit==4.2.0
